requirements.txt
nltk==3.7
numpy==1.23.5
pandas==1.5.3
PyYAML
scikit_learn==1.2.1
streamlit==1.26.0
streamlit_authenticator==0.2.3


